// apps/web/pharma-app/src/lib/estate.generated.ts
export type EstateTableKey =
  | 'mrd_reg.policies'
  | 'mrd_reg.sop_clause_links'
  | 'mrd_reg.sop_revisions'
  | 'mrd_reg.sops'
  | 'mrd_reg.standard_clauses'
  | 'mrd_reg.standards'
  | 'mrd_hcm.departments'
  | 'mrd_hcm.employees'
  | 'mrd_hcm.positions'
  | 'mrd_hcm.qualifications'
  | 'mrd_hcm.signature_authority'
  | 'mrd_hcm.training_curricula'
  | 'mrd_hcm.training_records'
  | 'mrd_erp.bill_of_materials'
  | 'mrd_erp.ingredients'
  | 'mrd_erp.lot_material_consumption'
  | 'mrd_erp.market_authorisations'
  | 'mrd_erp.material_lots'
  | 'mrd_erp.product_lots'
  | 'mrd_erp.products'
  | 'mrd_erp.suppliers'
  | 'mrd_mes.deviations'
  | 'mrd_mes.equipment'
  | 'mrd_mes.equipment_qualification'
  | 'mrd_mes.lines'
  | 'mrd_mes.process_parameters'
  | 'mrd_mes.process_steps'
  | 'mrd_mes.sites'
  | 'mrd_mes.work_orders'
  | 'mrd_qms.audits'
  | 'mrd_qms.batch_dispositions'
  | 'mrd_qms.capas'
  | 'mrd_qms.change_controls'
  | 'mrd_qms.complaints'
  | 'mrd_qms.lab_events'
  | 'mrd_qms.oos_investigations'
  | 'mrd_qms.qc_tests'
  | 'mrd_qms.spec_limits'
  | 'mrd_qms.specification_versions'
  | 'mrd_qms.specifications'
  | 'mrd_qms.test_methods'
  | 'mrd_tms.consignees'
  | 'mrd_tms.drivers'
  | 'mrd_tms.routes'
  | 'mrd_tms.shipment_lines'
  | 'mrd_tms.shipments'
  | 'mrd_tms.telematics_readings'
  | 'mrd_tms.trucks'
  | 'mrd_tms.warehouses'
  | 'mrd_kb.ask_history'
  | 'mrd_kb.document_chunks'
  | 'mrd_kb.document_relations'
  | 'mrd_kb.documents';